requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.3.0
chromadb==0.5.0
gradio==4.44.0
python-dotenv==1.0.0
openai==1.58.1
pypdf==5.0.0